Seperate HudPointsList.cs from PointsCheck.cs
